refacotr: integration GCP and local version
